* Add basic shellcode for mips-linux inside rasc.c * Fix segfault while using multiple color palettes * Add nop,metadata,header and printable palete colors * Fix build for arm and mips64 (get_reg int->u64)   - fix warns in get_tmpd_dir and endian_memcpy * More use of the color palette   - prompt color   - memory addresses   - for some headers   - byte types   - arch_aop in udis

+++ src/dbg/arch/arm.c
@@ -614,7 +614,7 @@ void free_bt(struct list_head *sf)
 	return;
 }
 
-int get_reg(char *reg)
+u64 get_reg(char *reg)
 {
 	elf_gregset_t regs;
 	int ret = ptrace (PTRACE_GETREGS, ps.tid, 0, &regs);
